+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. For commit 
  - Generating multiple HEIC previews in parallel fails #1470
  - ? in path not always encoded right #1447
  - fixed some condition that the halloween background doesn't load properly
+ - some comments not showing up on config viewer in settings
 
 ## v1.0.0